# openapi.yaml
openapi: "3.0.3"
info:
  title: Parqtel API
  description: |
    Ultra-lightweight SRE observability engine — streaming OpenTelemetry metrics, logs, and traces into compressed Parquet.
    Exposes Prometheus-compatible query API, Grafana SimpleJSON datasource, alerting engine, and pipeline management.
  version: "0.1.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Parqtel
    url: https://github.com/parqtel/parqtel-oss
servers:
  - url: http://localhost:9090
    description: Local development
  - url: http://localhost:8080
    description: Default production
...
components:
  schemas:
...
    IngestResponse:
      type: object
      properties:
        ingested:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
          example: "Validation error: Missing resource_metrics"
...
paths:
...
  /v1/logs:
    post:
      tags: [Ingestion]
      summary: Ingest logs (Protobuf)
      operationId: ingestLogsProto
      requestBody:
        required: true
        content:
          application/x-protobuf:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Logs ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/logs:
    get:
      tags: [Logs]
      summary: Query logs
      operationId: queryLogs
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: "Log filter expression (use * for all)"
        - name: start
          in: query
          required: true
          schema:
            type: number
          description: Start timestamp (Unix seconds)
        - name: end
          in: query
          required: true
          schema:
            type: number
          description: End timestamp (Unix seconds)
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: severity_min
          in: query
          schema:
            type: integer
          description: "Minimum severity number (9=INFO, 13=WARN, 17=ERROR, 21=FATAL)"
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search in log body
      responses:
        "200":
          description: Log query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      logs:
                        type: array
                        items:
                          type: object
                          properties:
                            timestamp_ns:
                              type: integer
                            severity_text:
                              type: string
                            body:
                              type: string
                            service_name:
                              type: string
                            attributes:
                              type: object
                      total_matched:
                        type: integer
                      truncated:
                        type: boolean
...